perf: add a brand new block_allocator
this allocator can bump the alloc speed with the help of memory
caching. TODO: finish the test of block_alloc() and block_free()

+++ include/fs.h
@@ -2,7 +2,7 @@
 #include <sys/types.h>
 
 #define FSMAGIC 0x636a6673
-#define BSIZE 4096
+#define BSIZE ((unsigned long)(4096))
 #define MAXOPBLOCKS 127
 #define NCACHE_BUF (MAXOPBLOCKS * 8)
 #define NLOG (MAXOPBLOCKS * 8)
